Added functionality for user registration,login and logout

+++ base/views.py
@@ -1,15 +1,63 @@
 from multiprocessing import context
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 
 from .models import Room, Topic
 from .forms import RoomForm
 
-# rooms = [
-#     {'id': 1, 'name': 'Lets learn python!'},
-#     {'id': 2, 'name': 'Design with me'},
-#     {'id': 3, 'name': 'Frontend developers'},
-#  ]
+
+
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated: #prevent user from accessing login page if already logged in
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'user does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'username or password does not exist')
+
+    context = {'page': page}    
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'an error occurred during registration')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -32,7 +80,7 @@ def room(request, pk):
     context = {'room': room}    
     return render(request, 'base/room.html', context)
    
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm() 
     if request.method == 'POST':
@@ -44,9 +92,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -59,9 +111,11 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
